# openapi.yaml
openapi: 3.0.3
info:
  title: AI Agent API
  description: |
    API pour le traitement automatisé des documents d'identité, permis de conduire et cartes grises avec intelligence artificielle.
    
    ## Fonctionnalités
    - Traitement OCR automatisé avec Azure Cognitive Services
    - Extraction intelligente des données structurées
    - Support multilingue (Français/Arabe)
    - Analytics et rapports en temps réel
    - Authentification JWT sécurisée
    
    ## Workflow Typique
    1. **Inscription/Connexion** : Créer un compte et récupérer un token JWT
    2. **Upload Documents** : Envoyer les images des documents (recto/verso)
    3. **Traitement IA** : L'API extrait et structure automatiquement les données
    4. **Récupération** : Accéder aux données structurées et aux analytics
    
    ## Support
    - Documentation interactive disponible sur `/docs/`
    - Exemples complets dans le README.md
    - Support technique : support@ai-agent.com
  version: 1.0.0
  contact:
    name: Support API AI Agent
    email: support@ai-agent.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Serveur de développement
  - url: https://api.ai-agent.com
    description: Serveur de production
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentification
      summary: Inscription d'un nouvel utilisateur
      description: Créer un nouveau compte utilisateur dans le système
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: Adresse email de l'utilisateur
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  description: Mot de passe (minimum 6 caractères)
                  example: motdepasse123
                full_name:
                  type: string
                  description: Nom complet de l'utilisateur
                  example: Jean Dupont
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentification
      summary: Connexion utilisateur
      description: Authentifier un utilisateur et récupérer un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: motdepasse123
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT Token d'authentification
                    example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Token obtenu via l'endpoint `/auth/login`.
        Format: `Bearer <your_jwt_token>`
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Message d'erreur détaillé
          example: Email and password required
...